requirements.txt
fastapi==0.88.0
uvicorn==0.22.0
langchain
langchain-nvidia-ai-endpoints
langchain-community
gradio==3.23.0
requests==2.31.0
rich==13.3.0
python-dotenv==1.0.0
faiss-cpu==1.7.3
numpy==1.23.5
pandas==1.5.3
